Helper to compute the confident homogeneity in a node.

Helper to compute the confident homogeneity in a node.

Usage

crit_node(valid_set, alpha, lambda)

Arguments

valid_set

(data.frame)
validation set. See get_valid_set() for details.

alpha

(proportion)
miscoverage rate.

lambda

(proportion)
balance between width and deviation.

Value

the confident homogeneity in the node.